[BoundsSafety] Remove `-funique-traps` flag and its implementation
The implementation was a bit hacky and interacted badly with the
hot-cold splitting pass. So it is going to be removed in favor of a
different implementation that's simpler.

I also investated why
`clang/test/BoundsSafety/CodeGen/unique-trap-blocks-O2.c` broke (rdar://150559639).
It turns outthat the "trap blocks" started being unintentionally split
by the hot-cold splitting pass because in upstream we started adding
branch profiling weights
(https://github.com/llvm/llvm-project/pull/134310) which the hot-cold
splitting pass used as a signal to split the "trap blocks" into their
own function (by-passing the logic I originally added to prevent the
splitting). While we could fix this it would be very fragile and given
that we are removing the implementation anyway it doesn't matter
anymore.

rdar://158088410