[LLDB][BoundsSafety] Add `InstrumentationRuntime::MatchAllModules`
This adds a virtual method that allows `InstrumentationRuntime`
sub-classes to match against all modules rather than just a library
that matches a particular regex. When the implementation returns
true `GetPatternForRuntimeLibrary()` is ignored and all modules
are iterated over. The default implementation returns false which
was the previous behavior which uses `GetPatternForRuntimeLibrary()`
to only match a particular runtime library.

The intended use case here is for implementing an
`InstrumentationRuntime` where the runtime library of interest can have
multiple implementations and whose name is not known ahead of time.

The concrete use case here is for a `InstrumentationRuntime` plugin
for implementations of the `-fbounds-safety` soft-trap runtime which
can have multiple different implementations.

An alternative to this would be for the `GetPatternForRuntimeLibrary()`
function to return a regex that matches everything. While that
technically works it seems wasteful to execute a regex for every loop
iteration.

rdar://FIXME