[Clang][LLDB] Refactor trap reason demangling out of LLDB and into Clang
This patch refactors the trap reason demangling logic in
`lldb_private::VerboseTrapFrameRecognizer::RecognizeFrame` into a new
public function `clang::CodeGen::DemangleTrapReasonInDebugInfo`.

There are two reasons for doing this:

1. In a future patch the logic for demangling needs to be used somewhere
   else in LLDB and thus the logic needs refactoring to avoid
   duplicating code.
2. The logic for demangling shouldn't really be in LLDB anyway because
   it's a Clang implementation detail and thus the logic really belongs
   inside Clang.

Unit tests have been added for the new function. Writing these tests
revealed there's an odd case that we currently allow where the category
and message strings are empty in `__builtin_verbose_trap`. We should
actually completely disallow that but that's out-of-scope for this patch
and fixing this is tracked by #165981.

This patch isn't entirely NFC because:

* The VerboseTrapFrameRecognizer previously recognized
  `__builtin_verbose_trap("", "")` and now it doesn't. A test case for
  this isn't included because the intention is to completely remove the
  ability to compile this in #165981.
* The LLDB logging output is a little different now. The previous code
  was a little too defensive and during the refactor it became apparent
  that some of the error paths that were previously being handled where
  unreachable so they handling them has been removed.